+++ src/convert/iso88591utf8codecvt.cpp
@@ -55,7 +55,7 @@ int iso88591utf8codecvt::do_length(const std::mbstate_t&,
     ++from_next;
   } // while
 
-  return (from_next-from);
+  return static_cast<int>(from_next-from);
 } // do_length
 
 // end of file
