+++ src/convert/utf16leucs2codecvt.cpp
@@ -1,6 +1,7 @@
 //---------------------------------------------------------------------------
 // $Id$
 //---------------------------------------------------------------------------
+#include <algorithm>
 #include <convert/utf16leucs2codecvt.hpp>
 #ifndef ARABICA_NO_WCHAR_T
 #include <convert/impl/ucs2_utf16.hpp>
@@ -46,7 +47,7 @@ int utf16leucs2codecvt::do_length(const std::mbstate_t&,
                                 const char* end,
                                 size_t max) const
 {
-  return std::max<int>((end-from), max/2);
+  return std::max<int>(static_cast<int>(end-from), static_cast<int>(max/2));
 } // do_length
 
 #endif
