+++ src/convert/utf8ucs2codecvt.cpp
@@ -77,7 +77,7 @@ int utf8ucs2codecvt::do_length(const std::mbstate_t&,
     }
   } // while
 
-  return (from_next-from);
+  return static_cast<int>(from_next-from);
 } // do_length
 
 #endif
